scripts/python/requirements.txt
paho-mqtt==1.1
protobuf<4.21.0
